+++ src/scripts/waypoints_list.js
@@ -13,9 +13,37 @@ const WaypointsList = {
             return resolve();
           }
 
+          /* damned infinite loop
           if (waited >= timeoutInSeconds * 1000) {
             return reject();
           }
+          */
+
+          wait(interval);
+        }, interval)
+      };
+
+      wait(100);
+    })
+  },
+
+  waitForElementHidden : (selector, timeoutInSeconds = 3) => {
+    return new Promise((resolve, reject) => {
+      let waited = 0;
+
+      const wait = (interval) => {
+        setTimeout(() => {
+          waited += interval;
+
+          if (!document.querySelector(selector)) {
+            return resolve();
+          }
+
+          /* damned infinite loop
+          if (waited >= timeoutInSeconds * 1000) {
+            return reject();
+          }
+          */
 
           wait(interval);
         }, interval)
@@ -27,7 +55,7 @@ const WaypointsList = {
 
   getRoutingSelectorHtml: () => {
     return `
-      <div id="routingSelector" class="SidebarRoutePoint">
+      <div id="routingSelector" class="MuiButtonBase-root MuiListItem-root MuiListItem-gutters MuiListItem-button">
         <span class="label">Set route: </span>
         <button class="btn" data-routing-type="fastest">
           <img class="icon" src="${chrome.extension.getURL("src/images/routing_types/fastest.svg")}" />
@@ -60,7 +88,7 @@ const WaypointsList = {
     document.querySelectorAll('.SidebarRoutePointCurrentRoutingProfile').forEach((button, index, items) => {
       setTimeout(() => {
         button.click();
-        const promise = WaypointsList.waitForElementReady(`.RoutingProfileSelection:not(#index-${index})`).then(() => {
+        WaypointsList.waitForElementReady(`.RoutingProfileSelection:not(#index-${index})`).then(() => {
           const RoutingProfileSelection = document.querySelector('.RoutingProfileSelection');
           RoutingProfileSelection.id = `index-${index}`;
           RoutingProfileSelection.querySelector(
@@ -72,17 +100,19 @@ const WaypointsList = {
   },
 
   addButtons: () => {
-    document.querySelector('.SidebarRouteAddViaInput').insertAdjacentHTML('beforeend', WaypointsList.getRoutingSelectorHtml());
+    console.log('add buttons')
+    document.querySelector('.SidebarRouteMore .MuiList-root').insertAdjacentHTML('afterbegin', WaypointsList.getRoutingSelectorHtml());
     document.querySelector('#routingSelector .label').addEventListener('click', WaypointsList.applyRoutin);
     document.querySelectorAll('#routingSelector .btn').forEach((button) => {
       button.addEventListener('click', WaypointsList.applyRouting)
     });
+
+    // wait for closing the panel
+    WaypointsList.waitForElementHidden('.SidebarRouteMore .MuiList-root').then(() => WaypointsList.set());
   },
 
   set: () => {
-    // eventlistener on "Plus" panel then wait for panel to display the routingSelector inside the panel
-    // this should fix the issue when the user display is track and come back to waypoints list
-
+    // wait for opening the "Plus" panel
     WaypointsList.waitForElementReady('.SidebarRouteMore .MuiList-root').then(() => WaypointsList.addButtons());
   }
 }